Moved enums into separate files
