+++ src/main/java/dataRepo/API.java
@@ -1,0 +1,7 @@
+package dataRepo;
+
+public enum API {
+	LEEWAY,
+	MARKETSTACK,
+	TWELVEDATA;
+}
