+++ src/main/java/dataRepo/FilterFlag.java
@@ -1,0 +1,26 @@
+package dataRepo;
+
+public enum FilterFlag {
+	STOCK(1 << 0),
+	ETF(1 << 1),
+	INDEX(1 << 2),
+	ALL((1 << 0) | (1 << 1) | (1 << 2));
+
+	private final int x;
+
+	FilterFlag(int x) {
+		this.x = x;
+	}
+
+	public int getVal() {
+		return x;
+	}
+
+	public static int getFilterVal(FilterFlag... flags) {
+		int val = 0;
+		for (FilterFlag flag : flags) {
+			val |= flag.getVal();
+		}
+		return val;
+	}
+}
